Change stuff

+++ sources/Game.hpp
@@ -8,7 +8,7 @@ namespace b_ball{
     class Game {
     private:
         Team _indoor, _outdoor;
-        size_t _indoor_points, _outdoor_points;
+        int _indoor_points, _outdoor_points;
     public:
         Game(Team& indoor, Team& outdoor);
         void roll_game_result();
